Filename handling overhaul
Matter now predominantly uses #list to refer to files,
rather than #sijil. #sijil and Sijils now refer to symbolic
representations of files, while #list and Tomes refer to a
list of the actual files. This should serve to allow Matter
to represent multiple arbitrary files much more easily, as
it won't be necessary for a #sijil to concretely represent
all of the desired files with a single glob pattern.